Hardware/robotics/huenit/external/huenit-robot-arm/requirements.txt
pyspacemouse
pyserial~=3.5